SIMPLE_BUGFIX: Fix actual value for repository count

+++ app/src/main/java/com/future/tailormade/command/order/impl/GetUserOrdersCommandImpl.java
@@ -42,7 +42,8 @@ public class GetUserOrdersCommandImpl implements GetUserOrdersCommand {
     private Mono<BasePagingResponse<GetOrdersResponse>> getUserOrdersCount(
             String userId, BasePagingResponse<GetOrdersResponse> pagingResponse) {
         return orderRepository.countAllByUserId(userId)
-                .map(count -> ResponseHelper.setPagingResponseTotalItem(pagingResponse, count));
+                .map(count -> ResponseHelper
+                        .setPagingResponseTotalItem(pagingResponse, count.intValue()));
     }
 
     private GetOrdersResponse createResponse(Order order) {
